gameovercontinuescreen redesign and animated

+++ Classes/game/uilayer/UIIngameAnimation.h
@@ -2,10 +2,10 @@
 #import "UIAnim.h"
 
 @interface UIIngameAnimation : CCSprite  {
-    int ct;
+    float ct;
 }
 
-@property(readwrite,assign) int ct;
+@property(readwrite,assign) float ct;
 -(void)update;
 
 @end
